requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
openpyxl>=3.0.0
xlrd>=2.0.0
python-dateutil>=2.8.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
